restructuring of HAL classes in progress.
